# src/routes.js
import { Router } from 'express';
import { comparePassword, hashPassword, signToken, authenticate, authorize } from './auth.js';
import { query } from './db.js';
import { employeeSchema, loginSchema, registerSchema, validate } from './validation.js';
export const router = Router();
const employeeSelect = `SELECT p.*, u.nama_unit, j.nama_jabatan, g.kode_golongan, g.nama_pangkat FROM pegawai p JOIN unit_kerja u ON u.id_unit=p.id_unit JOIN jabatan j ON j.id_jabatan=p.id_jabatan JOIN golongan g ON g.id_golongan=p.id_golongan`;
router.post('/auth/register', async (req, res, next) => { try { const parsed = validate(registerSchema, req.body); if (parsed.error) return res.status(400).json({ error: 'Validasi gagal.', details: parsed.error }); const { username, password } = parsed.data; const result = await query('INSERT INTO users (username, password_hash, role) VALUES ($1,$2,$3) RETURNING id_user, username, role', [username, await hashPassword(password), 'viewer']); return res.status(201).json(result.rows[0]); } catch (error) { return next(error); } });
router.post('/auth/login', async (req, res, next) => { try { const parsed = validate(loginSchema, req.body); if (parsed.error) return res.status(400).json({ error: 'Validasi gagal.', details: parsed.error }); const result = await query('SELECT * FROM users WHERE username=$1', [parsed.data.username]); if (!result.rows[0] || !(await comparePassword(parsed.data.password, result.rows[0].password_hash))) return res.status(401).json({ error: 'Username atau password salah.' }); return res.json({ token: signToken(result.rows[0]), user: { id_user: result.rows[0].id_user, username: result.rows[0].username, role: result.rows[0].role } }); } catch (error) { return next(error); } });
router.get('/pegawai', authenticate, async (req, res, next) => { try { const page = Math.max(Number(req.query.page) || 1, 1); const limit = Math.min(Math.max(Number(req.query.limit) || 10, 1), 100); const offset = (page - 1) * limit; const search = `%${String(req.query.search || '').trim()}%`; const rows = await query(`${employeeSelect} WHERE p.nama ILIKE $1 OR p.nip ILIKE $1 ORDER BY p.id_pegawai DESC LIMIT $2 OFFSET $3`, [search, limit, offset]); const total = await query('SELECT COUNT(*)::int AS total FROM pegawai WHERE nama ILIKE $1 OR nip ILIKE $1', [search]); return res.json({ data: rows.rows, pagination: { page, limit, total: total.rows[0].total } }); } catch (error) { return next(error); } });
router.get('/pegawai/:id', authenticate, async (req, res, next) => { try { const result = await query(`${employeeSelect} WHERE p.id_pegawai=$1`, [req.params.id]); if (!result.rows[0]) return res.status(404).json({ error: 'Pegawai tidak ditemukan.' }); return res.json(result.rows[0]); } catch (error) { return next(error); } });
router.post('/pegawai', authenticate, authorize('admin', 'editor'), async (req, res, next) => { try { const parsed = validate(employeeSchema, req.body); if (parsed.error) return res.status(400).json({ error: 'Validasi gagal.', details: parsed.error }); const v = parsed.data; const result = await query(`INSERT INTO pegawai (nip,nik,nama,jenis_kelamin,tempat_lahir,tanggal_lahir,tanggal_masuk,id_unit,id_jabatan,id_golongan,id_atasan,status_pegawai,batas_usia_pensiun) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13) RETURNING *`, Object.values(v)); return res.status(201).json(result.rows[0]); } catch (error) { return next(error); } });
router.put('/pegawai/:id', authenticate, authorize('admin', 'editor'), async (req, res, next) => { try { const parsed = validate(employeeSchema, req.body); if (parsed.error) return res.status(400).json({ error: 'Validasi gagal.', details: parsed.error }); const v = parsed.data; const result = await query(`UPDATE pegawai SET nip=$1,nik=$2,nama=$3,jenis_kelamin=$4,tempat_lahir=$5,tanggal_lahir=$6,tanggal_masuk=$7,id_unit=$8,id_jabatan=$9,id_golongan=$10,id_atasan=$11,status_pegawai=$12,batas_usia_pensiun=$13,updated_at=NOW() WHERE id_pegawai=$14 RETURNING *`, [...Object.values(v), req.params.id]); if (!result.rows[0]) return res.status(404).json({ error: 'Pegawai tidak ditemukan.' }); return res.json(result.rows[0]); } catch (error) { return next(error); } });
router.delete('/pegawai/:id', authenticate, authorize('admin'), async (req, res, next) => { try { const result = await query('DELETE FROM pegawai WHERE id_pegawai=$1', [req.params.id]); if (!result.rowCount) return res.status(404).json({ error: 'Pegawai tidak ditemukan.' }); return res.status(204).end(); } catch (error) { return next(error); } });